add setuptips to format description for Pypi
